Add files via upload
Leave management system consists of employee’s portal as well as admin’s portal. It is an intranet based application where employee can register, login and apply for leave and admin can either grant or deny the request. There are many types of leaves employee can apply for el, cl, scl, lwp. Database contains all the leave records of the faculty. This will reduce work for Head of departments. This application is an online application which makes more flexible to access information